backend/ECS/Train_module/requirements.txt
flask
mongoengine
flask-login
boto3
python-decouple
pymongo
mongoengine
pymongo[srv]
colorama 
dnspython 
argparse